added the same style for the buttons, and moved all css files to the styles folder
