+++ src/styles/ExerciseDetail.css
