+++ src/styles/progresstracker.css
@@ -11,6 +11,7 @@
 .description {
   background-color: rgba(0, 0, 0, 0.089);
   display: flex;
+  flex-direction: column;
   justify-content: center;
   width: 60dvw;
   font-family: "Roboto", sans-serif;
@@ -53,4 +54,8 @@ button {
   border-radius: 15px;
   margin-left: 10px;
   margin-bottom: 40px;
+}
+
+button:hover {
+  background-color: #b3072c;
 }
