+++ src/styles/watertracker.css
@@ -51,7 +51,7 @@
   
   
   .watertracker button:hover {
-    background-color: white; 
+    background-color: #b3072c; 
   }
   
   .inputs {
